// app/src/api/group/GroupClient.ts
export interface ExpenseView {
  id: string;
  author: string;
  description: string;
  amount: number;
  paid_by: string;
  split_between: string[];
  created_at: number;
}
// ...
export interface SettlementView {
  id: string;
  author: string;
  from: string;
  to: string;
  amount: number;
  created_at: number;
}